requirements.txt
blinker==1.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
h11==0.12.0
h2==4.0.0
hpack==4.0.0
hyperframe==6.0.0
idna==2.10
kaitaistruct==0.9
pyasn1==0.4.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
PySocks==1.7.1
python-dotenv==0.15.0
requests==2.25.1
selenium==3.141.0
selenium-wire==4.2.2
six==1.15.0
urllib3==1.26.3
wsproto==1.0.0
